Add pgSQL + SQLServer integration tests
