+++ StoredProcedureEFCore.ITest.PgSql/public/Stored Procedures/output_fixed_size.sql
@@ -1,0 +1,5 @@
+﻿CREATE OR REPLACE FUNCTION output_fixed_size(OUT fixed_size VARCHAR(255)) AS $$
+BEGIN
+    fixed_size = 'Jambon Beurre';
+END
+$$ LANGUAGE plpgsql;
