+++ StoredProcedureEFCore.ITest.PgSql/public/Stored Procedures/output_nullable.sql
@@ -1,0 +1,5 @@
+﻿CREATE OR REPLACE FUNCTION output_nullable(OUT nullable INT) AS $$
+BEGIN
+    nullable = null;
+END
+$$ LANGUAGE plpgsql;
