+++ StoredProcedureEFCore.ITest.SqlServer/dbo/Stored Procedures/empty.sql
@@ -1,0 +1,5 @@
+﻿CREATE PROCEDURE [dbo].[empty]
+AS
+BEGIN
+  SELECT * FROM table_1 WHERE id < 0;
+END
