+++ StoredProcedureEFCore.ITest.SqlServer/dbo/Stored Procedures/output_nullable.sql
@@ -1,0 +1,6 @@
+﻿CREATE PROCEDURE [dbo].[output_nullable]
+  @nullable INT OUTPUT
+AS
+BEGIN
+    SET @nullable = null
+END
