+++ StoredProcedureEFCore.ITest.SqlServer/dbo/Stored Procedures/select_param.sql
@@ -1,0 +1,6 @@
+﻿CREATE PROCEDURE [dbo].[select_param]
+    @n INT
+AS
+BEGIN
+    SELECT @n;
+END
