+++ StoredProcedureEFCore.ITest/ModelSlim.cs
@@ -1,0 +1,8 @@
+﻿namespace StoredProcedureEFCore.ITest
+{
+  class ModelSlim
+  {
+    public long Id { get; set; }
+    public string Name { get; set; }
+  }
+}
