--- StoredProcedureEFCore.Tests.PgSql.Database/public/Stored Procedures/ReturnBoolean.sql
@@ -1,5 +1,0 @@
-﻿CREATE OR REPLACE FUNCTION return_boolean(INOUT boolean_to_return BOOLEAN) RETURNS BOOLEAN
-LANGUAGE sql
-AS $$
-  SELECT boolean_to_return;
-$$;
