--- StoredProcedureEFCore.Tests.SqlServer.Database/dbo/Stored Procedures/Empty.sql
@@ -1,5 +1,0 @@
-﻿CREATE PROCEDURE [dbo].[Empty]
-AS
-BEGIN
-  SELECT * FROM Table_1 WHERE id < 0;
-END
