--- StoredProcedureEFCore.Tests.SqlServer.Database/dbo/Stored Procedures/ListNotAll.sql
@@ -1,8 +1,0 @@
-﻿CREATE PROCEDURE [dbo].[ListNotAll]
-	@limit bigint = 9223372036854775807
-AS
-BEGIN
-	SET NOCOUNT ON;
-
-	SELECT TOP(@limit) id, name, 5 AS extra_column FROM Table_1;
-END
