--- StoredProcedureEFCore.Tests.SqlServer.Database/dbo/Stored Procedures/SelectParam.sql
@@ -1,6 +1,0 @@
-﻿CREATE PROCEDURE [dbo].[SelectParam]
-	@n INT
-AS
-BEGIN
-	SELECT @n;
-END
