--- StoredProcedureEFCore.Tests.SqlServer/Model2.cs
@@ -1,8 +1,0 @@
-﻿namespace StoredProcedureEFCore.Tests.SqlServer
-{
-  class Model2
-  {
-    public long Id { get; set; }
-    public string Name { get; set; }
-  }
-}
