--- StoredProcedureEFCore.Tests.SqlServer/TestContext.cs
@@ -1,44 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace StoredProcedureEFCore.Tests.SqlServer
-{
-  public partial class TestContext : DbContext
-  {
-    public virtual DbSet<Table1> Table1 { get; set; }
-
-    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-    {
-      if (!optionsBuilder.IsConfigured)
-      {
-#warning To protect potentially sensitive information in your connection string, you should move it out of source code. See http://go.microsoft.com/fwlink/?LinkId=723263 for guidance on storing connection strings.
-        optionsBuilder.UseSqlServer(@"Server=.\SQLEXPRESS;Database=Test;Trusted_Connection=True;");
-      }
-    }
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-      modelBuilder.Entity<Table1>(entity =>
-      {
-        entity.ToTable("Table_1");
-
-        entity.Property(e => e.Id).HasColumnName("id");
-
-        entity.Property(e => e.Active).HasColumnName("active");
-
-        entity.Property(e => e.Date)
-                  .HasColumnName("date")
-                  .HasColumnType("datetime");
-
-        entity.Property(e => e.Name)
-                  .IsRequired()
-                  .HasColumnName("name")
-                  .HasMaxLength(255)
-                  .IsUnicode(false);
-
-        entity.Property(e => e.NameWithUnderscore)
-                  .HasColumnName("name_with_underscore")
-                  .HasDefaultValueSql("((0))");
-      });
-    }
-  }
-}
