+++ StoredProcedureEFCore.UTest/StoredProcedureEFCore.UTest.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+      <TargetFramework>netcoreapp3.1</TargetFramework>
+
+      <IsPackable>false</IsPackable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0"/>
+        <PackageReference Include="Moq" Version="4.13.1"/>
+        <PackageReference Include="Moq.DataReader" Version="0.1.1"/>
+        <PackageReference Include="NUnit3TestAdapter" Version="3.15.1"/>
+        <PackageReference Include="nunit" Version="3.12.0"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\StoredProcedureEFCore\StoredProcedureEFCore.csproj" />
+    </ItemGroup>
+
+</Project>
