+++ StoredProcedureEFCore/StoredProcedureEFCore.csproj
@@ -19,7 +19,7 @@
 
   <ItemGroup>
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-      <_Parameter1>$(AssemblyName).Tests</_Parameter1>
+      <_Parameter1>$(AssemblyName).UTest</_Parameter1>
     </AssemblyAttribute>
   </ItemGroup>
